[http_proxy_over_p2p] more tests, fix build
License: MIT
